pid tuning and parameters menu, more info on start screen

+++ src/eeprom.cpp
@@ -30,7 +30,7 @@ void EEPROM::init()
 {
     i2c.initI2C1Remapped();
     bool res = i2c.sendBytes(EEPROM_ADDRESS, nullptr, 0);
-    DEBUG_PRINT(DEBUG_DEBUG, "EEPROM detected=%u", res);
+    DEBUG_PRINT(DEBUG_DEBUG, "EEPROM detected=%u", (int)res);
 }
 
 void EEPROM::read()
@@ -38,7 +38,7 @@ void EEPROM::read()
     Data tmp;
     tmp.invalidate();
     bool result = eepromReadBytes(0, reinterpret_cast<uint8_t *>(&tmp), sizeof(tmp));
-    DEBUG_PRINT(DEBUG_WARNING, "read=%u magic=%08x version=%d sequence=%d", result, tmp.magic, tmp.version, tmp.sequence);
+    DEBUG_PRINT(DEBUG_WARNING, "read=%u magic=%08x version=%d sequence=%d", (int)result, tmp.magic, tmp.version, tmp.sequence);
 
     if (!result || tmp.magic != kMagic || tmp.version != kVersion) {
         resetDefaults();
@@ -70,7 +70,7 @@ bool EEPROM::write()
     if (!result) {
         data.sequence--;
     }
-    DEBUG_PRINT(DEBUG_ERROR, "write=%u magic=%08x version=%d sequence=%d", result, data.magic, data.version, data.sequence);
+    DEBUG_PRINT(DEBUG_ERROR, "write=%u magic=%08x version=%d sequence=%d", (int)result, data.magic, data.version, data.sequence);
 
     #if EEPROM_VALIDATE_WRITE
         tmp.invalidate();
